etl/requirements.txt
python-frontmatter==1.1.0
google-cloud-storage==2.19.0
